tests and figures

+++ apps/breakfast/bacon/cxActiveMessage/installDesktopVariablePower.sh
@@ -1,4 +1,26 @@
 #!/bin/bash
+retries=10
+retryTimeout=30
+locked=0
+for i in $(seq $retries)
+do
+  if [ -f install_lock ]
+  then
+    echo "install_lock present, retry in $retryTimeout s"
+    sleep $retryTimeout
+  else
+    echo "locking installation directory"
+    locked=1
+    touch install_lock
+    break
+  fi
+done
+if [ $locked -eq 0 ]
+then
+  echo "COULD NOT OBTAIN LOCK ON INSTALLATION DIRECTORY!" 1>&2
+  exit 1
+fi
+
 killall picocom
 autoRun=1
 programDelay=30
@@ -96,6 +118,7 @@ do
   if [ $varMatched -eq 0 ]
   then
     echo "Unrecognized option $1. Options: ${settingVars[@]}" 1>&2
+    rm install_lock
     exit 1
   fi
 done
@@ -108,6 +131,7 @@ done
 if [ "$testLabel" == "" ]
 then
   echo "No test label provided." 1>&2
+  rm install_lock
   exit 1
 fi
 
@@ -247,3 +271,5 @@ then
       install,$id bsl,ref,$ref
   done
 fi
+
+rm install_lock
